refactor old setup functions
